requirements.txt
pymilvus>=2.4.0                   # Milvus Python SDK
fastapi
uvicorn[standard]
mangum
python-dotenv
httpx
langchain-openai